+++ .test/quickPickProvider/alreadyExternalModuleImport.ts
@@ -1,0 +1,1 @@
+import bodyParse = require('body-parser');