+++ test/provider/ResolveCompletionItemProvider.test.ts
@@ -1,0 +1,27 @@
+import * as chai from 'chai';
+
+chai.should();
+
+// TODO: add when completion item provider is ready.
+
+// describe('ResolveCompletionItemProvider', () => {
+
+//     it('shoud resolve to null when less then 2 characters');
+
+//     it('shoud resolve to null if typing in a string');
+
+//     it('shoud resolve to null if after a "." (PropertyAccess)');
+
+//     it('shoud resolve to null if in a comment');
+
+//     it('shoud resolve to null if writing an import');
+
+//     it('shoud create a completion list');
+
+//     it('shoud add an insert text edit if import would be new');
+
+//     it('shoud add a replace text edit if import will be updated with new specifier');
+
+//     it('shoud add no text edit if import is already imported');
+
+// });