+++ .test/MyReactTemplate.tsx
@@ -1,0 +1,5 @@
+import * as React from "react";
+
+export const myComponent = (test: string) => {
+    return <div>Hello World: {test}</div>;
+}