+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.7.0]
 #### Added
 - More tests! :-) ([#8](https://github.com/buehler/typescript-hero/issues/8))
 - CodeCompletionProvider that autocompletes your symbols and adds the imports if necessary ([#5](https://github.com/buehler/typescript-hero/issues/5))
@@ -61,7 +63,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Various bugs in AST parsing
 
 
-[Unreleased]: https://github.com/buehler/typescript-hero/compare/v0.6.0...master
+[Unreleased]: https://github.com/buehler/typescript-hero/compare/v0.7.0...master
+[0.7.0]: https://github.com/buehler/typescript-hero/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/buehler/typescript-hero/compare/v0.5.4...v0.6.0
 [0.5.4]: https://github.com/buehler/typescript-hero/compare/v0.5.0...v0.5.4
 [0.5.0]: https://github.com/buehler/typescript-hero/compare/v0.4.0...v0.5.0